recipes-dev: Add configuration for empty password sudo
Let's make embedded development really easy.
The development image does not need to be secure.
The developer need to have access to the system.
The developer need to know that the system is not secure.
The developer need to know that this is not for production.
The developer has the responsibility to make the system secure.
The phobos-custom hardening mechanism will be used to make the system
secure.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>