recipes-common: hostname: Use machine as the hostname
The Torizon platform team had the weird idea of use the hostname as the
deviceID. The deviceID and the machine should have totally different
meanings. The deviceID is a unique identifier for the device, while the
machine is a generic name for the device. I mean, ok that the Toradex
devices uses the machine + serial id as the hostname, but this is not
the case for all devices. To workaround this, we will use the
machine name as the hostname, even though it is not the ideal.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>