recipes-bsp: u-boot: sl2619: fix fdt_addr_r
When booting with U-Boot on the SL2619 board, the fdt_addr_r environment
variable was incorrectly set to an address that could lead to
"corruptions" of the device tree blob in memory. This change updates the
fdt_addr_r to a safer address (0x10000000), which is less likely to
conflict with other memory regions used during the boot process.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>